+++ src/views/discover/components/DiscoverInvestisseur.vue
@@ -13,51 +13,48 @@
         <v-row>
           <v-col cols="3">
             <v-text-field
-              v-model="searchLastNameInput"
+              v-model="searchNameInput"
               :loading="loading"
-              label="Rechercher un investisseur par son nom"
-              placeholder="Saisissez un nom"
+              label="Search Projects"
+              placeholder="Start typing to Search"
               prepend-icon="mdi-magnify"
             />
           </v-col>
-          <!--v-col cols="3">
-            <v-text-field
-              v-model="searchFirstNameInput"
-              :loading="loading"
-              label="Rechercher un investisseur par son prénom"
-              placeholder="Saisissez un prénom"
-              prepend-icon="mdi-magnify"
-            />
-          </v-col-->
           <v-col cols="4">
             <v-select
-              v-model="job"
+              v-model="tags"
               :items="values"
-              label="Job"
+              label="Tags"
               multiple
               attach
-              @change="search()"
-              prepend-icon="mdi-alpha-j-circle "
+              prepend-icon="mdi-tag-multiple"
             />
           </v-col>
-          <v-col cols="4">
+          <v-col cols="2">
             <v-text-field
-              v-model="searchCompany"
+              v-model="searchBudgetMinInput"
               :loading="loading"
-              label="Entreprise"
-              placeholder="Saisissez le nom d'entreprise"
-              prepend-icon="mdi-domain"
-              suffix=""
+              label="Budget Min"
+              placeholder="0"
+              suffix="€"
+            />
+          </v-col>
+          <v-col cols="2">
+            <v-text-field
+              v-model="searchBudgetMaxInput"
+              :loading="loading"
+              label="Budget Max"
+              placeholder="999 999"
+              suffix="€"
             />
           </v-col>
         </v-row>
       </v-card-text>
-      <v-divider />
     </v-card>
     <v-row>
       <transition
-        v-for="investisseur in investisseursFiltered"
-        :key="investisseur.uid"
+        v-for="project in projectsFiltered"
+        :key="project.name"
         name="fade"
       >
         <v-col
@@ -65,35 +62,56 @@
         >
           <base-material-card
             class="v-card-profile"
-            :avatar="investisseur.photoURL"
+            :avatar="project.photoProjectURL"
           >
             <v-card-text class="text-center">
+              <v-row
+                justify="center"
+                class="mb-3"
+              >
+                <v-btn
+                  v-for="tag in project.tags"
+                  :key="tag"
+                  x-small
+                  :color="colorTag(tag)"
+                >
+                  {{ tag }}
+                </v-btn>
+              </v-row>
+
               <h6 class="display-1 mb-1 grey--text">
-                {{ investisseur.job }} chez {{ investisseur.company }}
+                {{ project.name }}
               </h6>
 
               <h4 class="display-2 font-weight-light mb-3 black--text">
-                <router-link v-bind:to="'/user/investor/' + investisseur.uid">
-                  {{ investisseur.nom }} {{ investisseur.prenom }}
-                </router-link>
+                {{ project.firstNameAuthor }} {{ project.lastNameAuthor }}
               </h4>
 
               <p class="font-weight-light grey--text">
-                {{ investisseur.abstract }}
+                {{ project.abstract }}
               </p>
 
-              <router-link v-bind:to="'/user/investor/' + investisseur.uid">
-                <v-btn
-                  color="primary"
-                  rounded
-                  class="mr-2"
-                  small
-                  @click="toDetailsInvestisseur(investisseur)"
-                >
-                  Regarder le profil
-                </v-btn>
-              </router-link>
-
+              <v-btn
+                color="primary"
+                rounded
+                class="mr-2"
+                small
+                :disabled="hasAccessProjects(project)"
+                @click="toDetailsProject(project)"
+              >
+                En savoir plus
+              </v-btn>
+              <v-btn
+                v-show="hasAccessProjects(project)"
+                color="secondary"
+                rounded
+                class="mr-0"
+                small
+                :disabled="hasDemandAccessProjects(project)"
+                @click="demandAccess(project)"
+              >
+                Demander l'acces
+              </v-btn>
             </v-card-text>
           </base-material-card>
         </v-col>
@@ -106,18 +124,17 @@
   import { mapMutations } from 'vuex'
 
   export default {
-    name: 'DiscoverCreateur',
+    name: 'DiscoverInvestisseur',
     data () {
       return {
-        searchLastNameInput: '',
-        // searchFirstNameInput: '',
-        searchCompany: '',
+        searchNameInput: '',
+        searchBudgetMinInput: '',
+        searchBudgetMaxInput: '',
         loading: false,
-        investisseurs: [],
-        investisseursFiltered: [],
-        jobFilter: [],
-        job: ['CTO', 'Développeur web', 'Autres'],
-        values: ['CTO', 'Développeur web', 'Autres'],
+        projects: [],
+        projectsFiltered: [],
+        tags: ['IT', 'web', 'crypto-currency', 'security'],
+        values: ['IT', 'web', 'crypto-currency', 'security'],
         history: [],
       }
     },
@@ -133,45 +150,58 @@
       },
     },
     watch: {
-      searchLastNameInput () {
-        this.search()
-      },
-      /* searchFirstNameInput () {
-        this.search()
-      }, */
-      searchCompany () {
-        this.search()
-      },
-      job () {
+      searchNameInput () {
         this.search()
       },
     },
     mounted () {
-      this.getInvestisseurs()
+      this.getProjects()
       this.history = this.userData.history
     },
-    // beforeDestroy () {
-    //  this.updateHistory()
-    //  this.setDataUser()
-    // },
+    beforeDestroy () {
+      this.updateHistory()
+      this.setDataUser()
+    },
     methods: {
-      toDetailsInvestisseur (investisseur) {
-        this.addToHistory(investisseur)
+      toDetailsProject (project) {
+        this.addToHistory(project)
         this.$router.push({
-          name: 'Investisseur',
-          params: { investisseur: investisseur },
+          name: 'Project',
+          params: { project: project },
         })
       },
-      async getInvestisseurs () {
-        await firestore.collection('users').get().then(projects => {
-          this.investisseurs = projects.docs.map(doc => {
+      hasAccessProjects (project) {
+        return !project.accessProject.includes(this.uid)
+      },
+      hasDemandAccessProjects (project) {
+        return project.accessDemand.includes(this.uid)
+      },
+      async demandAccess (project) {
+        project.accessDemand.push(this.uid)
+        await firestore.collection('projects').doc(project.uid).update({
+          accessDemand: project.accessDemand,
+        }).then(() => {
+          this.getProjects()
+        })
+      },
+      colorTag (tag) {
+        switch (tag) {
+          case 'web': return 'primary'
+          case 'crypto-currency': return 'secondary'
+          case 'security': return 'info'
+          case 'IT': return 'info'
+          default: return ''
+        }
+      },
+      async getProjects () {
+        await firestore.collection('projects').get().then(projects => {
+          this.projects = projects.docs.map(doc => {
             return {
               uid: doc.id,
               ...doc.data(),
             }
           })
-          this.investisseurs = this.investisseurs.filter(investisseur => investisseur.role === 'Investisseur')
-          this.investisseursFiltered = this.investisseurs
+          this.projectsFiltered = this.projects
         })
       },
       setDataUser () {
@@ -183,31 +213,22 @@
       },
       ...mapMutations({
         setUserData: 'SET_USER_DATA',
+        setAccessProjects: 'SET_ACCESS_PROJECTS',
       }),
       async updateHistory () {
         await firestore.collection('users').doc(this.uid).update({
           history: this.history,
         })
       },
-      addToHistory (investisseur) {
-        this.history.push(investisseur)
+      addToHistory (project) {
+        this.history.push(project)
       },
       search () {
-        // console.log(this.job.length)
         this.loading = true
-        if (this.searchLastNameInput === '' /* && this.searchFirstNameInput === '' */ && this.searchCompany === '' && this.job.length === 3) {
-          this.investisseursFiltered = this.investisseurs
-        } else if (this.job.length === 0) {
-          this.investisseursFiltered = []
+        if (this.searchInput !== '') {
+          this.projectsFiltered = this.projects.filter(item => item.name.includes(this.searchNameInput))
         } else {
-          this.investisseursFiltered = this.investisseurs.filter(item => item.nom.toLowerCase().includes(this.searchLastNameInput.toLowerCase()))
-          // this.investisseursFiltered = this.investisseurs.filter(item => item.prenom.toLowerCase().includes(this.searchFirstNameInput.toLowerCase()))
-          this.investisseursFiltered = this.investisseurs.filter(item => item.company.toLowerCase().includes(this.searchCompany.toLowerCase()))
-          this.investisseursFiltered.forEach(element => {
-            this.jobFilter.push(element.job.some(r => this.job.includes(r)) ? this.investisseursFiltered : this.investisseursFiltered.splice(this.investisseursFiltered.indexOf(element), 1))
-          })
-          // console.log(this.jobFilter)
-          this.jobFilter = []
+          this.projectsFiltered = this.projects
         }
         this.loading = false
       },
